Moved Dataset into field.py to resolve circular imports - maybe rename file. Also added extra arg handling to BaseField._log
